% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smoothLm.R
\name{calc_smooth_lm}
\alias{calc_smooth_lm}
\title{Calculate by smoothing and filling the columns, with linear models}
\usage{
calc_smooth_lm(
  data,
  based,
  value,
  max_fill = NULL,
  trailing = TRUE,
  name_as = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{data}{(data.frame) a \emph{ungrouped} data.frame storing all observations}

\item{based}{The column name of how the data should be arranged, in ascending order. This column must not contain NAs, must be unique, and must be "numericable".}

\item{value}{The column name of what values should be smoothed. This column must contain NAs.}

\item{max_fill}{(numeric) The maximum size of a NA-gap or \strong{special value} \code{NULL}, which ignores this argument.}

\item{trailing}{(logical) Remove all NAs at the head of prediction column? Default as \code{TRUE}.}

\item{name_as}{(character) Name of the new columns, the following are supported.
\itemize{
\item \code{NULL}: Create a new column as column name of \code{value} with prefix \code{slm_}
\item \verb{*del*}: Replaces the original column with the smoothed column
\item Any other character: Name of the column. May overwrite the original dataframe columns if \code{overwrite} == TRUE
}}

\item{overwrite}{(logical) Let the new column names to overwrite the original dataframe columns?
\itemize{
\item FALSE: Do not overwrite
\item TRUE: Do overwrite
}}
}
\value{
The same dataframe as \code{data}, with 1 additional column with column name defined in \code{name_as}.
}
\description{
Data smoothing is an important process in data analysis to fill up missing data, or to remove extreme values.
This function fills up missing data with linear model, predicted by the head and tail of the gap.
}
\examples{
calc_smooth_lm(data, x, y, trailing = F)
}
