% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_engine.R
\name{load_file.engine}
\alias{load_file.engine}
\title{Engine of the load_file function}
\usage{
load_file.engine(data_download, tmp_dir, threshold, ...)
}
\arguments{
\item{data_download}{(dataframe) Data to be downloaded, similar to \code{data} in \code{wx3::load_file}}

\item{tmp_dir}{(character) Temporary directory of the main \code{R} session}

\item{threshold}{(numeric) The minimum file size in bytes, similar to \code{threshold} in \code{wx3::load_file}}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\value{
NA
}
\description{
The engine to download files from a site, using parallel computation.
}
\keyword{internal}
