% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Download files from URL to Directory}
\usage{
load_file(
  data,
  worker = 0L,
  attempt = 5L,
  threshold = 1,
  list_fail = TRUE,
  title = "test",
  ...
)
}
\arguments{
\item{data}{(data.frame) Dataframe with character columns of \code{URL}, \code{DIR}, and \code{Info}. Read the description for more information.}

\item{worker}{(integer) Number of workers to download the data}

\item{attempt}{(integer) The maximum number of download attempts}

\item{threshold}{(numerical) The minimum file size in bytes}

\item{list_fail}{(logical) List failed to download items}

\item{title}{(character) Title of the downloaded items}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\value{
Downloaded files
}
\description{
Portal to download a list of files, using sequential and parallel programming. To initiate download process, input a dataframe with columns \code{DIR}, \code{URL}, \code{Info} representing the url, the destination directory and info of the data desired into \code{data}. The switch between sequential and parallel are determined by number of workers set.
If ...
\itemize{
\item \code{worker} < 0, Returns the dataframe
\item \code{worker} = 1, Initiate the download using sequential computation
\item \code{worker} > 1, Initiate the download using parallel computation
}

The download files may be incomplete, due to network issues. \code{threshold} determines the min. file size to be download. Any files with a smaller file size will be rejected. \code{attempt} determines the number of attempts to download a certain file. The next attempt(s) will only attempt to download the failed files in the current attempt. \code{list_fail} will list all failed download attempts.
}
