% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_SEQ.R
\name{load_file.SEQ}
\alias{load_file.SEQ}
\title{System tool: download files using sequential computation}
\usage{
load_file.SEQ(data, worker, attempt, threshold, list_fail, title, ...)
}
\arguments{
\item{data}{A \code{dataframe} with character columns of \code{URL}, \code{DIR}, and \code{Info}}

\item{attempt}{The maximum number of download attempts (\code{numerical})}

\item{threshold}{The minimum file size in bytes (\code{numerical})}

\item{list_fail}{List failed to download items (\code{logical})}

\item{title}{Title of the downloaded items (\code{character})}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\description{
Similar to \code{wx3::load_file}, but forced to have sequential computation.
}
\keyword{internal}
