% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_STM.R
\name{load_file.STMengine}
\alias{load_file.STMengine}
\title{System tool: Engine of the load_file function}
\usage{
load_file.STMengine(data_download, tmp_dir, threshold, ...)
}
\arguments{
\item{data_download}{Data to be downloaded, similar to \code{data} in \code{wx3::load_file}}

\item{tmp_dir}{Temporary directory of the main \code{R} session}

\item{threshold}{The minimum file size in bytes (\code{numerical}), similar to \code{threshold} in \code{wx3::load_file}}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\description{
The engine to download files from a site, using parallel computation.
}
\keyword{internal}
