% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fileset_SEQ.R
\name{load_fileset.SEQ}
\alias{load_fileset.SEQ}
\title{System tool: download filesets using sequential computation}
\usage{
load_fileset.SEQ(data, attempt, threshold, list_fail, title, ...)
}
\arguments{
\item{data}{\code{dataframe} with character columns of \code{URL}, \code{DIR}, and \code{Info}.}

\item{attempt}{(\code{numerical}) The maximum number of download attempts}

\item{threshold}{(\code{numerical}) The minimum file size in bytes}

\item{list_fail}{(\code{logical}) List failed to download items}

\item{title}{(\code{character}) Title of the downloaded items}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\description{
Similar to \code{wx3::load_fileset}, but forced to have sequential computation.
}
\keyword{internal}
