% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fileset_STM.R
\name{load_fileset.STM}
\alias{load_fileset.STM}
\title{System tool: download filesets using parallel computation}
\usage{
load_fileset.STM(data, worker, attempt, threshold, list_fail, title, ...)
}
\arguments{
\item{data}{A \code{dataframe} with character columns of \code{URL}, \code{DIR}, and \code{Info}}

\item{worker}{(\code{numerical}) Number of workers to download the data}

\item{attempt}{(\code{numerical}) The maximum number of download attempts}

\item{threshold}{(\code{numerical}) The minimum file size in bytes}

\item{list_fail}{(\code{logical}) List failed to download items}

\item{title}{(\code{character}) Title of the downloaded items}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\description{
Similar to \code{wx3::load_fileset}, but forced to have parallel computation.
}
\keyword{internal}
