% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fileset_STM.R
\name{load_fileset.STMengine}
\alias{load_fileset.STMengine}
\title{System tool: Engine of the load_fileset function}
\usage{
load_fileset.STMengine(data_download, tmp_dir, threshold, ...)
}
\arguments{
\item{data_download}{(\code{data.frame}) Data to be downloaded, similar to \code{data} in \code{wx3::load_fileset}}

\item{tmp_dir}{(\code{character}) Temporary directory of the main \code{R} session}

\item{threshold}{(\code{numerical}) The minimum file size in bytes, similar to \code{threshold} in \code{wx3::load_fileset}}

\item{...}{Additional parameters to be passed to \code{RCurl::getBinaryURL}}
}
\description{
The engine to download filesets from a site, using parallel computation.
}
\keyword{internal}
