% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Panel.R
\name{load_panelstart}
\alias{load_panelstart}
\alias{load_panelend}
\title{System tools: Print out the load panel}
\usage{
load_panelstart(title, time_start, data_1, data_n, data_num, storm = FALSE)

load_panelend(
  time_start,
  time_end,
  attempts,
  success,
  fail,
  list_fail,
  list_of_fail
)
}
\arguments{
\item{title}{Title of the downloaded items}

\item{time_start}{Time of the download process starts}

\item{data_1}{The first item to be downloaded}

\item{data_n}{The last item to be downloaded}

\item{data_num}{Number of items to download}

\item{time_end}{Time of the download process ends}

\item{attempts}{Total attempts made to download all the items}

\item{success}{Total number of success}

\item{fail}{Total number of fails}

\item{list_fail}{Should you list the failed-to-download items?}

\item{list_of_fail}{The list of failed to download items.}
}
\description{
System tools: Print out the load panel
}
\examples{
sys_ldhp_panelstart("test", Sys.time(), "1", "n", 999)
sys_ldhp_panelend(Sys.time(), Sys.time(), 999, 0, 9, TRUE, "fail")
}
\keyword{internal}
