% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-has_internet.R
\name{sys_hasInternet}
\alias{sys_hasInternet}
\title{Internet check}
\usage{
sys_hasInternet(test_host = NULL, attempt = 5L)
}
\arguments{
\item{test_host}{(character) Array of website to be tested for testing internet connection. Accepts following value.
\itemize{
\item NULL: Default list of website to be tested, which can be modified in \code{extdata}
\item Any other character: Website URL/ IP addresses.
For most purposes, please input the URL instead of IP address, as testing DNS server connectivity is also practical for most purposes.
}}

\item{attempt}{(integer) Attempts to be made per website listed in \code{test_host}}
}
\value{
logical value
\itemize{
\item TRUE: Has internet connection
\item FALSE: No internet connection
}
}
\description{
Test if there is internet access, based on network connect with websites. By default, the list of websites to be tested can be found in \code{sys_hasInternet.txt} in the package.

By default, the following website will be used for testing
\itemize{
\item google.com
\item bing.com
\item un.org/zh
}

Other website can also be used for testing, by inputing the address into the \code{test_host}
}
\examples{
sys_hasInternet()
}
\keyword{internal}
