% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-session_info.R
\name{sys_info}
\alias{sys_info}
\title{System tool: Return system info}
\usage{
sys_info()
}
\value{
A list of system information
}
\description{
System tool: Return system info
}
\examples{
wx3::sys_info()
}
\keyword{internal}
