% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-msg_warn.R
\name{sys_warn}
\alias{sys_warn}
\title{System tools: A slightly better warning messenger}
\usage{
sys_warn(message, x, arg = rlang::caller_arg(x), ...)
}
\arguments{
\item{message}{Message to be displayed, formatted via \code{cli::cli_bullets()}. For more information, please read \code{cli::cli_abort}.}

\item{x}{An argument name in the current function.}

\item{arg}{Argument to be traced. When referring to this in argument \code{message}, use \verb{\{.arg \{arg\}\}}.}

\item{...}{Additional items to consider in the message. Any items that uses glue (\code{{}}) must be passed through here.}
}
\value{
A warning message
}
\description{
Custom \code{cli} formating can be used to refer different stuff. This function is based on \code{cli::cli_bullets} for formatting. Note that when writing functions, \strong{do not} use pipe (\verb{\%>\%}, \verb{|>}) before this function.

Supports inline-markup themes. For more information on those custom themes, please refer to \code{wx3:::sys_msg_theme} and \code{cli::cli_div}.
}
\examples{
wx3:::sys_warn(message = "test")
}
\keyword{internal}
